requirements.txt
psutil
entrypoints
jsonschema>=3.0
future
matplotlib
openpyxl
tensorboard>=2.9.1
google-auth-oauthlib<1.1,>=0.5
tensorboard-data-server<0.8.0,>=0.7.0
werkzeug>=1.0.1
requests==2.25.1
sacremoses
nltk==3.7
datasets==2.0.0
transformers~=4.19.2
scipy~=1.7.3
pandas~=1.3.5
numpy<2,>=1.22.1
librosa~=0.8.0
tqdm~=4.62.3
omegaconf~=2.0.6
pillow~=9.0.1
cython~=0.29.28
soundfile~=0.10.3.post1
toml==0.10.2
inflect==5.6.0
importlib-resources